+++ rust/tableau_summary/src/twb/datasource/object_graph.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use roxmltree::Node;
 use tracing::info;
-use crate::twb::data_source::connection::{Expression, Join, Relation, Table, Union};
+use crate::twb::datasource::connection::{Expression, Join, Relation, Table, Union};
 use crate::twb::xml::XmlExt;
 
 #[derive(Serialize, Deserialize, Default, PartialEq, Clone, Debug)]
